move geo and vector modules into utils folder
