+++ src/index.js
@@ -8,8 +8,8 @@ import Scene from './scene/scene';
 import version from './utils/version';
 import log from './utils/log';
 import Utils from './utils/utils';
-import Geo from './geo';
-import Vector from './vector';
+import Geo from './utils/geo';
+import Vector from './utils/vector';
 import DataSource from './sources/data_source';
 import GLSL from './gl/glsl';
 import ShaderProgram from './gl/shader_program';
