+++ src/leaflet_layer.js
@@ -1,6 +1,6 @@
 import Thread from './utils/thread';
 import Scene from './scene/scene';
-import Geo from './geo';
+import Geo from './utils/geo';
 import debounce from './utils/debounce';
 import {mergeDebugSettings} from './utils/debug_settings';
