+++ src/scene/scene_worker.js
@@ -6,7 +6,7 @@ import debugSettings, {mergeDebugSettings} from '../utils/debug_settings';
 import log from '../utils/log';
 import WorkerBroker from '../utils/worker_broker'; // jshint ignore:line
 import Tile from '../tile/tile';
-import Geo from '../geo';
+import Geo from '../utils/geo';
 import DataSource from '../sources/data_source';
 import '../sources/sources';
 import FeatureSelection from '../selection/selection';
