+++ src/sources/raster.js
@@ -1,6 +1,6 @@
 import DataSource, {NetworkTileSource} from './data_source';
 import {TileID} from '../tile/tile_id';
-import Geo from '../geo';
+import Geo from '../utils/geo';
 import Texture from '../gl/texture';
 import Utils from '../utils/utils';
 import hashString from '../utils/hash';
