+++ src/styles/lines/lines.js
@@ -8,7 +8,7 @@ import Texture from '../../gl/texture';
 import VertexLayout from '../../gl/vertex_layout';
 import {buildPolylines} from '../../builders/polylines';
 import renderDashArray from './dasharray';
-import Geo from '../../geo';
+import Geo from '../../utils/geo';
 
 import WorkerBroker from '../../utils/worker_broker';
 import hashString from '../../utils/hash';
