+++ src/tile/tile.js
@@ -1,7 +1,7 @@
 import log from '../utils/log';
 import Utils from '../utils/utils';
 import mergeObjects from '../utils/merge';
-import Geo from '../geo';
+import Geo from '../utils/geo';
 import {TileID} from './tile_id';
 import {addLayerDebugEntry} from '../styles/style';
 import StyleParser from '../styles/style_parser';
