+++ src/tile/tile_manager.js
@@ -1,7 +1,7 @@
 import Tile from './tile';
 import {TileID} from './tile_id';
 import TilePyramid from './tile_pyramid';
-import Geo from '../geo';
+import Geo from '../utils/geo';
 import mainThreadLabelCollisionPass from '../labels/main_pass';
 import log from '../utils/log';
 import WorkerBroker from '../utils/worker_broker';
