+++ src/utils/obb.js
@@ -1,4 +1,4 @@
-import Vector from '../vector';
+import Vector from './vector';
 
 export default class OBB {
