+++ src/utils/vector.js
